requirements.txt
requests==2.31.0
aiohttp==3.9.1
pandas==2.1.4
numpy==1.26.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
scipy==1.11.4
tenacity==8.2.3
pytz==2024.1
dash==2.14.2
plotly==5.18.0
pytest==7.4.3
pytest-asyncio==0.21.1
